+++ tests/cli/test_run_benchmark_with_timeouts.py
@@ -21,8 +21,6 @@ def test_benchmark_run_no_timeout_does_not_fail(tmp_path):
         result = omni.call(
             [
                 "run",
-                "benchmark",
-                "--benchmark",
                 (tmp_path / timeout_benchmark).as_posix(),
             ],
             cwd=tmp_path,
@@ -41,8 +39,6 @@ def test_benchmark_run_with_timeout_less_than_task_fails(tmp_path):
         result = omni.call(
             [
                 "run",
-                "benchmark",
-                "--benchmark",
                 (tmp_path / timeout_benchmark).as_posix(),
                 "--task-timeout",
                 "1s",
@@ -61,8 +57,6 @@ def test_benchmark_run_high_timeout(tmp_path):
         result = omni.call(
             [
                 "run",
-                "benchmark",
-                "--benchmark",
                 (tmp_path / timeout_benchmark).as_posix(),
                 "--task-timeout",
                 "5s",