+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](https://semver.org/) and [Conventi
 - refactor: rename `info` cli group to `describe`
 - refactor: change suborganization of `storage` and rename into `remote files`, `remote version` and `remote policy`
 - refactor: change behaviour of `--local-storage` flag and rename into `--use-remote-storage`
+- refactor(cli)!: simplify and unify `ob run` command - benchmark file is now a positional argument and module functionality is integrated via optional flags (`ob run benchmark.yaml` for benchmarks, `ob run benchmark.yaml --module MODULE_ID --input-dir DIR` for modules)
 - bug: do not fail if numeric values given in commit, params or version (#187)
 - bug: Use -- separator for passing delimiting extra arguments passing to snakemake
 - bug: do not hardcode /tmp in run_module (#173)